fix for case where item is not found

+++ src/OpenCOVER/OpenVRUI/base/coMenu.cpp
@@ -440,7 +440,7 @@ coMenuItemVector coMenu::getAllItems()
     return res;
 }
 
-size_t coMenu::index(const coMenuItem *item)
+int coMenu::index(const coMenuItem *item)
 {
     coMenuItemVector allItems = getAllItems();
     size_t i = 0;
